Require loaders to be used within a GraphQL::Batch.batch block
This is needed to avoid global state leaking between requests/tests.
